pos_neg_lsvm_no_reweighting_for_aciotn

+++ src/fr/durandt/jstruct/latent/lssvm/ranking/ROCAreaLSSVMCuttingPlane1Slack.java
@@ -60,7 +60,7 @@ public abstract class ROCAreaLSSVMCuttingPlane1Slack<X,H> extends LSSVMCuttingPl
 		}
 
 		// Divide by the number of relevant examples * number of irrelevant examples
-		double c = (double)(x.getNpos()*x.getNneg());
+		double c = x.getNpos()*x.getNneg();
 		for(int d=0; d<dim; d++) {
 			psi[d] /= c;
 		}
