+++ src/fr/durandt/jstruct/latent/mantra/cvpr15/multiclass/FastMulticlassMantraCVPRCuttingPlane1SlackBagImage.java
@@ -191,7 +191,7 @@ public class FastMulticlassMantraCVPRCuttingPlane1SlackBagImage extends FastMult
 		double accuracy = 0;
 		int nb = 0;
 		for(STrainingSample<LatentRepresentation<BagImage, Integer>,Integer> ts : l){
-			int ypredict = (int) prediction(ts.input);
+			int ypredict = prediction(ts.input);
 			if(ts.output == ypredict){	
 				nb++;
 			}
