+++ src/fr/durandt/jstruct/latent/mantra/iccv15/multiclass/FastMulticlassMantraCuttingPlane1SlackBagImage.java
@@ -192,7 +192,7 @@ public class FastMulticlassMantraCuttingPlane1SlackBagImage extends FastMulticla
 		double accuracy = 0;
 		int nb = 0;
 		for(STrainingSample<LatentRepresentation<BagImage, Integer>,Integer> ts : l){
-			int ypredict = (int) prediction(ts.input);
+			int ypredict = prediction(ts.input);
 			if(ts.output == ypredict){	
 				nb++;
 			}
