+++ src/fr/durandt/jstruct/latent/mantra/iccv15/multiclass/MulticlassMantraCuttingPlane1SlackBagImageRegion.java
@@ -203,7 +203,7 @@ public class MulticlassMantraCuttingPlane1SlackBagImageRegion extends MantraCutt
 		double accuracy = 0;
 		int nb = 0;
 		for(STrainingSample<LatentRepresentation<BagImageRegion, Integer>,Integer> ts : l){
-			int ypredict = (int) prediction(ts.input);
+			int ypredict = prediction(ts.input);
 			if(ts.output == ypredict){	
 				nb++;
 			}
