+++ src/fr/durandt/jstruct/latent/mantra/iccv15/multiclass/MulticlassMantraPegasosBagImageRegion.java
@@ -203,7 +203,7 @@ public class MulticlassMantraPegasosBagImageRegion extends MantraPegasos<BagImag
 		double accuracy = 0;
 		int nb = 0;
 		for(STrainingSample<LatentRepresentation<BagImageRegion, Integer>,Integer> ts : l){
-			int ypredict = (int) prediction(ts.input);
+			int ypredict = prediction(ts.input);
 			if(ts.output == ypredict){	
 				nb++;
 			}
