+++ src/fr/durandt/jstruct/variable/BagLabel.java
@@ -66,7 +66,7 @@ public class BagLabel extends Bag {
 				proportion += 1;
 			}
 		}
-		labelProportion = proportion/(double)labels.size();
+		labelProportion = proportion/labels.size();
 	}
 
 	/**
